% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOP.R
\name{avg_probs}
\alias{avg_probs}
\title{Average probabilities}
\usage{
avg_probs(df_cdfs)
}
\arguments{
\item{df_cdfs}{data.frame containing trimmed ecdf values (output from \code{trim_cdf()})}
}
\value{
vector of probabilities for each value (specified in \code{evaluate_cdf()} argument \code{vals})
}
\description{
Calculate average of trimmed probabilities for each value. Step 4 in LOP calculation.
}
