% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOP.R
\name{calculate_aggregate_LOP}
\alias{calculate_aggregate_LOP}
\title{Implement probability averaging}
\usage{
calculate_aggregate_LOP(dat, keep_vals, ret_quantiles)
}
\arguments{
\item{dat}{data.frame containing cdfs to be combined. \code{dat} should contain three columns, named \code{id}, \code{quantile}, and \code{value}.
\code{id} contains identifiers to distinguish between different individual cdfs.
\code{quantile} contains quantiles (between 0 and 1) and \code{value} contains the cdf value for that id and quantile.}

\item{keep_vals}{ADD}

\item{ret_quantiles}{A matrix with length(x) rows, and n columns where n is the number of distributions to combine}
}
\value{
A data.frame with \code{quantile} and \code{value} of probability averaged aggregate.
}
\description{
Given a set of cumulative distribution functions  (assume for now: defined on same values),
combine using probability averaging (also called linear opinion pool). #TODO CITE
This method calculates the (weighted) average of quantiles at a given value
}
\examples{
dat <- expand.grid(id = c("A", "B"),
                   quantile = seq(0,1,0.01))
dat$value <- ifelse(dat$id == "A", qnorm(dat$quantile), qnorm(dat$quantile, 0,2))
calculate_aggregate_vin(dat, seq(0,1,0.05))

}
