% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vincent.R
\name{calculate_aggregate_vin}
\alias{calculate_aggregate_vin}
\title{Implement quantile averaging}
\usage{
calculate_aggregate_vin(dat, ret_quantiles)
}
\arguments{
\item{dat}{data.frame containing cdfs to be combined. \code{dat} should contain three columns, named \code{id}, \code{quantile}, and \code{value}.
\code{id} contains identifiers to distinguish between different individual cdfs.
\code{quantile} contains quantiles (between 0 and 1) and \code{value} contains the cdf value for that id and quantile.}

\item{ret_quantiles}{A matrix with length(x) rows, and n columns where n is the number of distributions to combine}
}
\value{
A data.frame with \code{quantile} and \code{value} of quantile averaged aggregate.
}
\description{
Given a set of cumulative distribution functions  (assume for now: defined on same values),
combine using quantile averaging (also called Vincent average).
This method calculates the (weighted) average of values at a given quantile
}
\examples{
dat <- expand.grid(id = c("A", "B"),
                   quantile = seq(0,1,0.01))
dat$value <- ifelse(dat$id == "A", qnorm(dat$quantile), qnorm(dat$quantile, 0,2))
calculate_aggregate_vin(dat, seq(0,1,0.05))

}
