% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementAggregation.R
\name{calculate_single_aggregate}
\alias{calculate_single_aggregate}
\title{export}
\usage{
calculate_single_aggregate(
  data,
  id_var,
  method,
  ret_quantiles,
  trim = "none",
  n_trim = NA
)
}
\description{
export
}
