% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{cdf_mean}
\alias{cdf_mean}
\title{Approximate mean of ecdf}
\usage{
cdf_mean(cdf)
}
\arguments{
\item{cdf}{data.frame containing \code{quantile} and \code{value} columns}
}
\description{
Given an empirical cdf, use triangle rule to estimate mean of distribution.
}
