% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfTrim.R
\name{cdf_trim}
\alias{cdf_trim}
\title{Implement cdf trimming - TO UPDATE}
\usage{
cdf_trim(data, trim_type, n_trim, avg_dir)
}
\arguments{
\item{data}{data.frame containing columns for \code{id}, \code{quantile}, \code{value}}

\item{trim_type}{"interior" for omitting central distributions or
"exterior" for omitting outermost distributions}

\item{n_trim}{integer, number of ids to trim}

\item{avg_dir}{string specifying whether averaging is "vincent" or "LOP"}
}
\value{
data.frame containing original \code{data} argument plus additional column for \code{weight},
which specifies which values are excluded (i.e., given 0 weight)
}
\description{
Cdf trimming removes a set number of innermost/outermost probabilities.
at each value before averaging \strong{ADD REF}.
}
\details{
\code{keep_vals} should contain integer values between 1 and \code{n_id},
the number of unique ids. Then, at each value, \code{trim_cdf()} will order probabilities and
keep only those that defined by \code{keep_vals}. For example, if \code{n_id = 5} and
\code{keep_vals = 2:4}, \code{trim_cdf()} will return the second, third and fourth ranked probability,
excluding the smallest and larges probabilities from the average.
}
