% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOP.R
\name{create_interp_fns}
\alias{create_interp_fns}
\title{Interpolate empirical cdfs}
\usage{
create_interp_fns(dat)
}
\arguments{
\item{dat}{data.frame containing cdfs to be combined. \code{dat} should contain three columns, named \code{id}, \code{quantile}, and \code{value}.
\code{id} contains identifiers to distinguish between different individual cdfs.
\code{quantile} contains quantiles (between 0 and 1) and \code{value} contains the cdf value for that id and quantile.}
}
\value{
a list of functions defining linearly interpolated cdfs (input: value, output: quantile),
where list elements are named by \code{id}.
}
\description{
Perform linear interpolation on set of empirical cdfs. Step 1 in Linear Opinion Pool calculation.
}
\details{
This function performs completeness checks for inputted cdfs. To be included,
input cdfs are excluded if they meet one of the following conditions:
\enumerate{
\item all quantiles are equal to a single value
\item any value contains NA
\item decreasing values for increasing quantiles
}
}
