% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOP.R
\name{evaluate_cdf}
\alias{evaluate_cdf}
\title{Evaluate unique cdfs for standard values}
\usage{
evaluate_cdf(vals, interp_functions)
}
\arguments{
\item{vals}{standard values over which to calculate probability for each unique cdf}

\item{interp_functions}{linearly interpolated cdf functions (output from \code{create_interp_fns()})}
}
\value{
data.frame with columns \code{id}, \code{quantile}, and \code{value}
}
\description{
Calculate probability for standard values across all unique cdfs. Step 2 in LOP calculation
}
