% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vincent.R
\name{evaluate_qf}
\alias{evaluate_qf}
\title{function to interpolate to consistent quantiles}
\usage{
evaluate_qf(quantile, value, id, ret_quants)
}
\description{
function to interpolate to consistent quantiles
}
