% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{keep_vals}
\alias{keep_vals}
\title{Identify values to trim}
\usage{
keep_vals(trim_type, n_trim, n_ids, avg_dir, method)
}
\arguments{
\item{trim_type}{"interior" for omitting central distributions or
"exterior" for omitting outermost distributions}

\item{n_trim}{integer, number of ids to trim}

\item{n_ids}{integer, number of unique ids being aggregated}

\item{avg_dir}{"LOP" for probability averaging or "vincent" for quantile averaging}

\item{method}{"cdf" or "mean" for the trimming method REF}
}
\value{
list containing two elements: (1) vector of values to keep;
(2) number of ids trimmed
}
\description{
Based on trimming direction, method, and number, return values to trim. TO DO: Lower/upper
}
\details{
In some cases,the number of models, the number of ids to trim
and the trimming type are not compatible. In these cases, \code{keep_vals()}
modifies n_trim. The following rules must be met for compatibility:
\enumerate{
\item For exterior trimming, \code{n_trim \leq floor(n_models/2)}.
\item For interior trimming, \code{n_trim} and \code{n_models} must have same
parity (i.e., both are even or both are odd).
TODO: ADD CHECKS/WARNINGS HERE, EXAMPLES
}
}
