% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineTrimVals.R
\name{keep_vals_exterior}
\alias{keep_vals_exterior}
\title{Setup exterior trimming}
\usage{
keep_vals_exterior(n_trim, n_ids)
}
\arguments{
\item{n_trim}{integer, number of ids to trim}

\item{n_ids}{integer, number of unique ids being aggregated}
}
\value{
list containing two elements: (1) vector of values to keep;
(2) number of ids trimmed
}
\description{
For exterior trimming, remove \code{n_trim/2} lowest and highest values.
}
\examples{
keep_vals_exterior(2, 4)
keep_vals_exterior(4, 5)
}
