% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineTrimVals.R
\name{keep_vals_interior}
\alias{keep_vals_interior}
\title{Setup interior trimming}
\usage{
keep_vals_interior(n_trim, n_ids)
}
\arguments{
\item{n_trim}{integer, number of ids to trim}

\item{n_ids}{integer, number of unique ids being aggregated}
}
\value{
list containing two elements: (1) vector of values to keep;
(2) number of ids trimmed
}
\description{
For interior trimming, remove innermost \code{n_trim} values.
}
\details{
Interior trimming requires (1) \code{n_trim < n_ids} and
(2) \code{n_trim} and \code{n_ids} have the same parity
(i.e., both are even or both are odd). If not,
\code{n_trim} is increased to \code{n_trim + 1}.
}
\examples{
\dontrun{
keep_vals_interior(2, 4)
keep_vals_interior(1, 4)
}
}
