% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOP.R
\name{LOP_distribution}
\alias{LOP_distribution}
\title{Back to quantile function UPDATE}
\usage{
LOP_distribution(gdf_cdf, vals, ret_q)
}
\arguments{
\item{gdf_cdf}{vector of probabilities for each value (output from \code{avg_probs()})}

\item{vals}{vector of unique values (specified in \code{evaluate_cdf()} argument \code{vals})}

\item{ret_q}{vector of quantiles to return values from aggregate LOP distribution}
}
\description{
Return LOP aggregate distribution for specified quantiles. Step 5 in LOP calculation.
}
