% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTrim.R
\name{mean_trim}
\alias{mean_trim}
\title{Implement mean trimming - TODO: UPDATE}
\usage{
mean_trim(data, trim_type, n_trim)
}
\arguments{
\item{trim_type}{"interior" for omitting central distributions or
"exterior" for omitting outermost distributions}

\item{n_trim}{integer, number of ids to trim}

\item{avg_dir}{"LOP" for probability averaging or "vincent" for quantile averaging}
}
\value{
data.frame containing \code{quantile} and \code{value} of aggregate distribution,
plus columns specifying \code{direction}, \code{method}, \code{trim_type}, and \code{n_trim}.
}
\description{
Trim entire distribution based on mean value ADD REF.
}
