% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{mean_trim_ids}
\alias{mean_trim_ids}
\title{Identify trimmed ids}
\usage{
mean_trim_ids(dat, keep_vals)
}
\arguments{
\item{dat}{data.frame containing cdfs to be combined. \code{dat} should contain three columns, named \code{id}, \code{quantile}, and \code{value}.
\code{id} contains identifiers to distinguish between different individual cdfs.
\code{quantile} contains quantiles (between 0 and 1) and \code{value} contains the cdf value for that id and quantile.}

\item{keep_vals}{vector specifying which probabilities to keep at a given value. See details.}
}
\description{
Given empirical cdfs, identify which ids to trim based on mean of distribution.
}
\details{
\code{keep_vals} should contain integer values between 1 and \code{n_id},
the number of unique ids. Then, at each value, \code{trim_cdf()} will order probabilities and
keep only those that defined by \code{keep_vals}. For example, if \code{n_id = 5} and
\code{keep_vals = 2:4}, \code{trim_cdf()} will return the second, third and fourth ranked probability,
excluding the smallest and larges probabilities from the average.
}
