% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MMODS}
\alias{MMODS}
\title{Projections from Multiple Models for Outbreak Decision Support (MMODS) study}
\format{
A data frame with 1700 rows and 3 variables:
\describe{
\item{id}{id to differentiate projections from distinct models}
\item{quantile}{projection quantile from 0 to 1}
\item{value}{projected cumulative number of deaths for that quantile}
}
}
\source{
\url{https://github.com/MMODS-org/Elicitation-1}
}
\usage{
MMODS
}
\description{
This data includes projections from 17 distinct models elicited during the first
round of the MMODS study. These projections are of cumulative deaths from
May-November 2020 in a US county of approximately 100,000 individuals that
implements a strict stay-at-home order for the duration of the projection period.
Details on the elicitation and the data can be found in Shea et al. 2020.
}
\keyword{datasets}
