% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MMODS_obs}
\alias{MMODS_obs}
\title{Observations from Multiple Models for Outbreak Decision Support (MMODS) study}
\format{
A data frame with 84 rows and 5 variables:
\describe{
\item{FIPS}{FIPS code for each county}
\item{County}{name of county}
\item{State}{state in which county is}
\item{Population}{popualtion size of county}
\item{cumu_deaths}{number of COVID-19 deaths recorded in county through November 15, 2020}
}
}
\source{
\url{https://github.com/MMODS-org/Elicitation-1}
}
\usage{
MMODS_obs
}
\description{
This data includes observations from 84 US counties with population between
90,000 and 110,000 individuals that implemneted a strict stay-at-home orders
between May and November 2020. These observations represent cumulative deaths
in these counties.
Details on the elicitation and the data can be found in Shea et al. 2020.
}
\keyword{datasets}
