% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOP.R
\name{trim_cdf}
\alias{trim_cdf}
\title{Implement cdf trimming}
\usage{
trim_cdf(df_long, keep_vals)
}
\arguments{
\item{df_long}{data.frame containing probabilities across standardized values for each unique cdf
(output from \code{evaluate_cdf})}

\item{keep_vals}{vector specifying which probabilities to keep at a given value. See details.}
}
\value{
data.frame with columns \code{id}, \code{quantile}, and \code{value} containing trimmed ecdfs
}
\description{
Cdf trimming removes a set number of innermost/outermost probabilities
at each value before averaging \strong{ADD REF}. Step 3 in LOP calculation.
}
\details{
\code{keep_vals} should contain integer values between 1 and \code{n_id},
the number of unique ids. Then, at each value, \code{trim_cdf()} will order probabilities and
keep only those that defined by \code{keep_vals}. For example, if \code{n_id = 5} and
\code{keep_vals = 2:4}, \code{trim_cdf()} will return the second, third and fourth ranked probability,
excluding the smallest and larges probabilities from the average.
}
