% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tuesdata}
\alias{tuesdata}
\title{Weather Forecast Accuracy}
\format{
A tt_data object with 3 elements
\describe{
\item{weather_forecasts}{a data frame with 651968 rows and 10 variables}
\item{cities}{a data frame with 236 rows and 11 variables}
\item{outlook_meanings}{a data frame with 23 rows and 2 variables}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/tree/master/data/2022/2022-12-20}
}
\usage{
tuesdata
}
\description{
Weather Forecast Accuracy
}
\keyword{datasets}
