% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vincent.R
\name{vincent}
\alias{vincent}
\title{Implement quantile averaging}
\usage{
vincent(data, ret_quantiles, weight_fn = equal_weights, ...)
}
\value{
A data.frame with \code{quantile} and \code{value} of quantile averaged aggregate.
}
\description{
Given a set of cumulative distribution functions  (assume for now: defined on same values),
combine using quantile averaging (also called Vincent average).
This method calculates the (weighted) average of values at a given quantile
}
\examples{
dat <- expand.grid(id = c("A", "B"),
                   quantile = seq(0,1,0.01))
dat$value <- ifelse(dat$id == "A", qnorm(dat$quantile), qnorm(dat$quantile, 0,2))
calculate_aggregate_vin(dat, seq(0,1,0.05))

}
