% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementAggregation.R
\name{aggregate_cdfs}
\alias{aggregate_cdfs}
\title{aggregate a set of cdfs}
\usage{
aggregate_cdfs(
  data,
  id_var,
  group_by,
  method,
  ret_quantiles,
  weighting_scheme = "equal",
  ...
)
}
\arguments{
\item{data}{data.frame that contains multiple cdfs, grouped by \code{by} column.
Specify cdf with \code{quantile} and \code{value} columns,}

\item{id_var}{string containing the name of the column that identifies unique cdfs}

\item{group_by}{vector containing the names of the columns to create unique aggregates for}

\item{method}{character name of the method for aggregation (see details).}

\item{weighting_scheme}{string to indicate how to weight in the aggregate (see details).}

\item{n_trim}{integer denoting the number of models to trim}
}
\value{
TBD
}
\description{
Given a data.frame containing cdfs , return a single aggregate
cdf using specified method.
}
\details{
Methods include:
\enumerate{
\item "LOP" - simple probability averaging, also called Linear Opinion Pool.
\item "vincent" - simple quantile averaging, also called Vincent average.
}

Weighting schemes include:
\enumerate{
\item "equal" - equal weighting of all models and values
\item "user_defined" - user supplies weights applied to each model
(additional input \code{weights} that is a data.frame containing \code{id} and \code{weight} columns)
\item trimming - "cdf_interior", "cdf_exterior",
"mean_interior", "mean_exterior", following REF (additional inputs ...)
}
}
