% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementAggregation.R
\name{aggregate_cdfs}
\alias{aggregate_cdfs}
\title{aggregate a set of cdfs}
\usage{
aggregate_cdfs(
  data,
  id_var,
  group_by,
  method,
  ret_quantiles,
  trim = "none",
  n_trim = NA
)
}
\arguments{
\item{data}{data.frame that contains multiple cdfs, grouped by \code{by} column.
Specify cdf with \code{quantile} and \code{value} columns,}

\item{id_var}{string containing the name of the column that identifies unique cdfs}

\item{group_by}{vector containing the names of the columns to create unique aggregates for}

\item{method}{character name of the method for aggregation. See details for methods.}

\item{trim}{string to indicate trimming method (see Details)}

\item{n_trim}{integer denoting the number of models to trim}
}
\value{
TBD
}
\description{
Given a data.frame containing cdfs , return a single aggregate
cdf using specified method.
}
\details{
Methods include
\enumerate{
\item "LOP" - simple probability averaging, also called Linear Opinion Pool.
\item "vincent" - simple quantile averaging, also called Vincent average.
Trim inputs should be one of the following: "cdf_interior", "cdf_exterior",
"mean_interior", "mean_exterior", or "none" following REF.
}
}
