% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implementAggregation.R
\name{calculate_single_aggregate}
\alias{calculate_single_aggregate}
\title{export}
\usage{
calculate_single_aggregate(
  quant,
  val,
  id,
  method,
  ret_quantiles,
  weighting_scheme = "equal",
  ...
)
}
\description{
export
}
