% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdfTrim.R
\name{cdf_trim}
\alias{cdf_trim}
\title{Implement cdf trimming - TO UPDATE}
\usage{
cdf_trim(data, trim_type, n_trim, avg_dir)
}
\arguments{
\item{data}{data.frame containing ids, quantiles, and values}

\item{trim_type}{string specifying "interior" or "exterior"}

\item{n_trim}{integer specifying the number of values to trim}

\item{avg_dir}{string indicating the direction of the averaging/trimming; use "LOP" or "vin".}
}
\value{
data.frame with columns \code{id}, \code{quantile}, and \code{value} containing trimmed ecdfs

data.frame containing \code{quantile} and \code{value} of aggregate distribution,
plus columns specifying \code{direction}, \code{method}, \code{trim_type}, and \code{n_trim}.
TO DO: MAKE THIS CONSISTENT
}
\description{
Cdf trimming removes a set number of innermost/outermost probabilities.
at each value before averaging \strong{ADD REF}.
}
\details{
\code{keep_vals} should contain integer values between 1 and \code{n_id},
the number of unique ids. Then, at each value, \code{trim_cdf()} will order probabilities and
keep only those that defined by \code{keep_vals}. For example, if \code{n_id = 5} and
\code{keep_vals = 2:4}, \code{trim_cdf()} will return the second, third and fourth ranked probability,
excluding the smallest and larges probabilities from the average.
}
