% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOP.R
\name{LOP}
\alias{LOP}
\title{Implement probability averaging}
\usage{
LOP(data, ret_quantiles, weight_fn = equal_weights, ...)
}
\arguments{
\item{data}{data.frame containing \code{quantile}, \code{value} columns}

\item{ret_quantiles}{vector of quantiles to return specifying the aggregate distribution}
}
\description{
Given a set of cumulative distribution functions  (assume for now: defined on same values),
combine using probability averaging (also called linear opinion pool). #TODO CITE
This method calculates the (weighted) average of quantiles at a given value
}
