% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{mean_trim}
\alias{mean_trim}
\title{Implement mean trimming}
\usage{
mean_trim(dat, n_trim, trim_type, avg_dir, ret_quantiles)
}
\arguments{
\item{dat}{data.frame containing cdfs to be combined. \code{dat} should contain three columns, named \code{id}, \code{quantile}, and \code{value}.
\code{id} contains identifiers to distinguish between different individual cdfs.
\code{quantile} contains quantiles (between 0 and 1) and \code{value} contains the cdf value for that id and quantile.}

\item{n_trim}{integer, number of ids to trim}

\item{trim_type}{"interior" for omitting central distributions or
"exterior" for omitting outermost distributions}

\item{avg_dir}{"LOP" for probability averaging or "vincent" for quantile averaging}

\item{ret_quantiles}{A matrix with length(x) rows, and n columns where n is the number of distributions to combine}
}
\value{
data.frame containing \code{quantile} and \code{value} of aggregate distribution,
plus columns specifying \code{direction}, \code{method}, \code{trim_type}, and \code{n_trim}.
TO DO: MAKE THIS NAMING CONSISTENT
}
\description{
Trim entire distribution based on mean value ADD REF.
}
