% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTrim.R
\name{mean_trim}
\alias{mean_trim}
\title{Implement mean trimming - TODO: UPDATE}
\usage{
mean_trim(data, trim_type, n_trim)
}
\arguments{
\item{data}{data.frame containing columns for \code{id}, \code{quantile}, \code{value}}

\item{trim_type}{"interior" for omitting central distributions or
"exterior" for omitting outermost distributions}

\item{n_trim}{integer, number of ids to trim}
}
\value{
data.frame containing original \code{data} argument plus additional column for \code{weight},
which specifies which cdfs are excluded (i.e., given 0 weight)
}
\description{
Trim entire distribution based on mean value ADD REF.
}
