% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vincent.R
\name{vincent}
\alias{vincent}
\title{Implement quantile averaging}
\usage{
vincent(quantile, value, id, ret_quantiles, weight_fn = equal_weights, ...)
}
\arguments{
\item{quantile}{vector containing quantiles for all cdfs to be aggregated}

\item{value}{vector containing values for all cdfs to be aggregated}

\item{id}{vector containing unqiue ids to distinguish cdf to be aggregated}

\item{ret_quantiles}{vector of quantiles to return specifying the aggregate distribution}
}
\value{
vector of values for corresponding \code{ret_quantiles} of aggregate distribution
}
\description{
Given a set of cumulative distribution functions  (assume for now: defined on same values),
combine using quantile averaging (also called Vincent average).
This method calculates the (weighted) average of values at a given quantile
}
\examples{
dat <- expand.grid(id = c("A", "B"),
                   quantile = seq(0,1,0.01))
dat$value <- ifelse(dat$id == "A", qnorm(dat$quantile), qnorm(dat$quantile, 0,2))
vincent(dat$quantile, dat$value, dat$id, seq(0,1,0.05))

}
